requirements.txt
Django==2.0
django-jet==1.0.8
djangorestframework==3.11.0
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
